Presale

+++ contracts/Airdrop.sol
@@ -7,19 +7,19 @@ import '@openzeppelin/contracts/security/ReentrancyGuard.sol';
 import '@openzeppelin/contracts/token/ERC20/IERC20.sol';
 
 contract Airdrop is Ownable, ReentrancyGuard {
-    uint256 public totalClaimed;
-    uint256 public amountToClaim;
-    uint256 public claimCount;
-    uint256 public timeOut;
+    uint public totalClaimed;
+    uint public amountToClaim;
+    uint public claimCount;
+    uint public timeOut;
     address burnAddress = 0x000000000000000000000000000000000000dEaD;
     mapping (address => bool) public addressReceived;
     IERC20 public token;
-    event eventClaimed(address sender, uint256 amount);
-    event eventBurnRemainingTokens(uint256 amount);
-    event eventSetAmountToClaim(uint256 amount);
+    event eventClaimed(address sender, uint amount);
+    event eventBurnRemainingTokens(uint amount);
+    event eventSetAmountToClaim(uint amount);
     event eventSetTokenAddress(address amount);
 
-    constructor(address _token, uint256 _amountToClaim) {
+    constructor(address _token, uint _amountToClaim) {
         token = IERC20(_token);
         amountToClaim = _amountToClaim;
     }
@@ -35,7 +35,7 @@ contract Airdrop is Ownable, ReentrancyGuard {
         emit eventClaimed(msg.sender, amountToClaim);
     }
 
-    function start(uint256 timeSeconds) public nonReentrant onlyOwner {
+    function start(uint timeSeconds) public nonReentrant onlyOwner {
         require(timeOut == 0, 'start: Airdrop has already started');
         timeOut = block.timestamp + timeSeconds;
     }
@@ -43,12 +43,12 @@ contract Airdrop is Ownable, ReentrancyGuard {
     function burnRemainingTokens() public { // to be fair anyone can burn remaining tokens when airdrop is over
         require(timeOut != 0, 'burnRemainingTokens: Airdrop has not started yet');
         require(timeOut < block.timestamp, 'burnRemainingTokens: Airdrop has not ended yet');
-        uint256 remaining = getRemainingTokens();
+        uint remaining = getRemainingTokens();
         require(token.transfer(burnAddress, remaining));
         emit eventBurnRemainingTokens(remaining);
     }
 
-    function getRemainingTokens() public view returns (uint256) {
+    function getRemainingTokens() public view returns (uint) {
         return token.balanceOf(address(this));
     }
 
@@ -57,7 +57,7 @@ contract Airdrop is Ownable, ReentrancyGuard {
         emit eventSetTokenAddress(_tokenAddress);
     }
 
-    function setAmountToClaim(uint256 _amount) public onlyOwner {
+    function setAmountToClaim(uint _amount) public onlyOwner {
         amountToClaim = _amount;
         emit eventSetAmountToClaim(_amount);
     }
